kernel: make the ip layer more link layer agnostic
This moves the responsibility of creating a link layer packet from
the ip layer to the low level device driver. This doesn't cause any
additional overhead, as the ethernet frame is constructed directly
in the devices send buffer.

+++ kernel/include/kernel/net/udp.h
@@ -17,9 +17,9 @@ struct udp_packet {
     uint8_t payload[0];
 } __attribute__((packed));
 
-ssize_t net_send_udp_through_socket(struct socket *socket, const void *buf, size_t len, const struct sockaddr *dest);
-ssize_t net_send_udp(struct network_interface *interface, struct ip_v4_address dest, uint16_t source_port, uint16_t dest_port, uint16_t len,
-                     const void *buf);
+int net_send_udp_through_socket(struct socket *socket, const void *buf, size_t len, const struct sockaddr *dest);
+int net_send_udp(struct network_interface *interface, struct ip_v4_address dest, uint16_t source_port, uint16_t dest_port, uint16_t len,
+                 const void *buf);
 void net_udp_recieve(const struct udp_packet *packet, size_t len);
 void net_init_udp_packet(struct udp_packet *packet, uint16_t source_port, uint16_t dest_port, uint16_t len, const void *buf);
